% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqalign-package.R
\docType{package}
\name{ggseqalign-package}
\alias{ggseqalign}
\alias{ggseqalign-package}
\title{ggseqalign: Minimal Visualization of Sequence Alignments}
\description{
Simple visualizations of alignments of DNA or AA sequences as well as arbitrary strings. Compatible with Biostrings and ggplot2. The plots are fully customizable using ggplot2 modifiers such as theme().
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simeross/ggseqalign}
  \item Report bugs at \url{https://github.com/simeross/ggseqalign/issues}
}

}
\author{
\strong{Maintainer}: Simeon Lim Rossmann \email{simeon.rossmann@nmbu.no} (\href{https://orcid.org/0000-0003-0435-8221}{ORCID})

}
\keyword{internal}
