% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment_table.R
\name{alignment_table}
\alias{alignment_table}
\title{Alignment Table}
\usage{
alignment_table(query = XStringSet, subject = XStringSet)
}
\arguments{
\item{query}{An object of class XStringSet containing the query sequences.}

\item{subject}{An object of class XStringSet containing the subject sequence.}
}
\value{
A list containing tibbles with information on mismatches and indels.
}
\description{
Generate a table of mismatches and indels between one or many query sequences and a subject sequence.
}
\examples{
query_seq <- DNAStringSet(c("ACCGTACCTGG", "ACCTTGG"))
subject_seq <- DNAStringSet("ACCGTACCGGG")
x <- alignment_table(query_seq, subject_seq)
x

}
