% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_alignment.R
\name{plot_sequence_alignment}
\alias{plot_sequence_alignment}
\title{Plot sequence alignment}
\usage{
plot_sequence_alignment(alignment_tbl = alignment_table(query, subject))
}
\arguments{
\item{alignment_tbl}{An alignment table containing query and subject information for sequence alignment.}
}
\value{
A ggplot object representing the sequence alignment plot.
}
\description{
This function generates a sequence alignment plot using ggplot2 based on the input alignment table.
}
\examples{
q <- (c("boo", "fizzbuzz"))
s <- ("boofizz")

alignment <- alignment_table(q, s)
plot_sequence_alignment(alignment_tbl = alignment)

# Provide names for query and subject to label the y-axis
names(q) <- c("Seq1", "Seq2")
names(s) <- "reference"

alignment <- alignment_table(q, q)
plot_sequence_alignment(alignment_tbl = alignment)



}
